[CDEC-470] Catch IOExceptions in `productionReporter`
Reporting is currently complex and brittle, and one failure mode occurs
when the network connection goes down, and reporting tries to report,
but dies due to an IOException because the network is down.

Installing a simple `catchIO` handler around `reportNode` should catch
such exceptions and report their occurence in the log.